(refactor) Move generateCoordinates to separate file
